requirements.txt
altair==4.2.2
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
htmlmin==0.1.12
idna==3.7
ImageHash==4.3.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.1.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
multimethod==1.12
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
phik==0.12.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.36.0
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visions==0.7.4
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
torch==2.3.1
torchvision==0.18.1
transformers==4.42.4
